+++ app/models/censor_rule.rb
@@ -117,7 +117,10 @@ class CensorRule < ApplicationRecord
   end
 
   def make_regexp(encoding)
-    Regexp.new(encoded_text(encoding), Regexp::MULTILINE)
+    ::Warning.with_raised_warnings do
+      Regexp.new(encoded_text(encoding), Regexp::MULTILINE)
+    end
+  rescue RaisedWarning => e
+    raise RegexpError, e.message.split('warning: ').last.chomp
   end
-
 end